big progress on base mechanics

+++ Assets/MoneyManager.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class MoneyManager : MonoBehaviour
 {
     public static MoneyManager manager;
-    public int money = 0;
+    public float money = 0;
 
     private void Start()
     {
@@ -16,4 +16,5 @@ public class MoneyManager : MonoBehaviour
     {
         money += amount;
     }
+
 }
